requirements.txt
asgiref==3.11.0
attrs==25.4.0
colorama==0.4.6
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==39.0.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
hypothesis==6.98.16
pytest-cov==5.0.0
pytest-html==4.1.1
pytest-xdist==3.6.1
pytest-sugar==1.0.0
pytest-rich==0.2.0
pytest-timeout==2.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0

gunicorn==23.0.0


whitenoise==6.8.2

